backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
minio
pydantic
requests
